Area.java

+++ Area.java
@@ -3,7 +3,7 @@ public class Area{
 	public static void main(String args[]){
 		Scanner sc = new Scanner(System.in);
 		int r;
-		int Ar;
+		double Ar;
 		System.out.println("Enter the Radius:");
 		r = sc.nextInt();
 		Ar = (3.14)*r*r;
